requirements.txt
# AI Video Generator - Requirements
# Install: pip install -r requirements.txt

# Core
requests>=2.28.0
python-dotenv>=1.0.0

# AI/LLM
google-generativeai>=0.3.0

# Text-to-Speech
gTTS>=2.3.0
pydub>=0.25.0

# Video Processing
moviepy>=1.0.3
Pillow>=10.0.0
numpy>=1.24.0

# Web Scraping (for RSS/Reddit)
beautifulsoup4>=4.12.0
feedparser>=6.0.0

# Utilities
tqdm>=4.65.0
colorama>=0.4.6

# YouTube API
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0

# Optional but recommended
# google-cloud-texttospeech>=2.14.0  # For Google Cloud TTS (requires setup)
